pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "orchestra-dienstplan"
version = "0.1.0"
description = "TVK-konformer Dienstplan-Generator für die Sächsische Bläserphilharmonie"
requires-python = ">=3.9"
dependencies = [
    "openpyxl>=3.1",
    "pdfplumber>=0.9",
    "pyyaml>=6.0",
    "click>=8.0",
    "pydantic>=2.0",
    "python-dateutil>=2.9",
    "holidays>=0.40",
    "python-docx>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov>=4.0"]

[project.scripts]
dienstplan = "dienstplan.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
